responsiveness finished
