+++ app/globals.css
@@ -54,6 +54,9 @@
 .main-background {
   background-color: #19171A;
 }
+.slick-track {
+  
+}
 
 
 @layer utilities {
